accukonx_social_network/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.4
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
et-xmlfile==1.1.0
Events==0.5
exceptiongroup==1.2.1
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.4.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.4.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-groq==0.0.1
langchain-openai==0.0.6
langsmith==0.1.51
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.21.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.23.6
openpyxl==3.1.2
opensearch-py==2.6.0
orjson==3.10.1
packaging==23.2
PAM==0.4.2
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==4.25.3
psycopg2==2.9.9
ptyprocess==0.7.0
pyarrow==16.0.0
pycairo==1.20.1
pycountry==23.12.11
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0b1
Pygments==2.17.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.35.0
regex==2024.4.28
reportlab==3.6.8
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.2
streamlit==1.31.1
systemd-python==234
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
validators==0.28.1
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==4.0.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
yarl==1.9.4
zipp==1.0.0
